+++ src/index.ts
@@ -15,7 +15,8 @@ export {
 export { BuiltinLogger } from "./builtin-logger";
 export { Middleware } from "./middleware";
 export {
-  createResultHandler,
+  AbstractResultHandler,
+  ResultHandler,
   defaultResultHandler,
   arrayResultHandler,
 } from "./result-handler";
@@ -47,7 +48,6 @@ export type { FlatObject } from "./common-helpers";
 export type { Method } from "./method";
 export type { IOSchema } from "./io-schema";
 export type { CommonConfig, AppConfig, ServerConfig } from "./config-type";
-export type { ResultHandlerDefinition } from "./result-handler";
 export type { ApiResponse } from "./api-response";
 export type {
   BasicSecurity,